photobot_src/requirements3.txt
requests
urllib3
idna
certifi
chardet
sqlalchemy
gps3
configparser
scapy
lanscan
RPi.GPIO
pytz
onvif_zeep
ruamel.yaml